+++ src/app/API/fake.api/user.api.js
@@ -166,7 +166,6 @@ const users = [
         bookmark: false
     }
 ];
-
 if (!localStorage.getItem("users")) {
     localStorage.setItem("users", JSON.stringify(users));
 }
@@ -177,14 +176,13 @@ const fetchAll = () =>
             resolve(JSON.parse(localStorage.getItem("users")));
         }, 2000);
     });
-
 const update = (id, data) =>
     new Promise((resolve) => {
         const users = JSON.parse(localStorage.getItem("users"));
-        const userId = users.findIndex((u) => u._id === id);
-        users[userId] = { ...users[userId], ...data };
+        const userIndex = users.findIndex((u) => u._id === id);
+        users[userIndex] = { ...users[userIndex], ...data };
         localStorage.setItem("users", JSON.stringify(users));
-        resolve(users[userId]);
+        resolve(users[userIndex]);
     });
 
 const getById = (id) =>
@@ -197,7 +195,6 @@ const getById = (id) =>
             );
         }, 1000);
     });
-
 export default {
     fetchAll,
     getById,
