+++ src/app/API/index.js
@@ -1,7 +1,6 @@
 import users from "./fake.api/user.api";
 import professions from "./fake.api/professions.api";
 import qualities from "./fake.api/qualities.api";
-
 const API = {
     users,
     professions,
