+++ src/app/components/page/userPage/index.js
@@ -1,2 +1,8 @@
 import UserPage from "./userPage";
-export default UserPage;
+import UserPageEdit from "./userPageEdit";
+const userPageOne = {
+    UserPage,
+    UserPageEdit
+};
+
+export default userPageOne;
