+++ src/app/components/page/usersListPage/usersListPage.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import Pagination from "../../common/pagination";
 import { paginate } from "../../../utils/paginate";
-import PropTypes from "prop-types";
+// import PropTypes from "prop-types";
 import api from "../../../api";
 import GroupList from "../../common/groupList";
 import SearchStatus from "../../ui/searchStatus";
@@ -140,8 +140,8 @@ const UsersListPage = () => {
     return <h1>loading...</h1>;
 };
 
-UsersListPage.propTypes = {
-    users: PropTypes.array
-};
+// UsersListPage.propTypes = {
+//     users: PropTypes.array
+// };
 
 export default UsersListPage;
